+++ src/namdesktop/model/ProjectTemplate.java
@@ -1,0 +1,9 @@
+package namdesktop.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.List;
+
+public record ProjectTemplate(
+        @JsonProperty("name")     String name,
+        @JsonProperty("children") List<TemplateNode> children) {}